pyproject.toml
[project]
name = "notary-assistant"
version = "0.1.0"
description = "Notary Assistant for Stamp and Certify Co"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "google-generativeai>=0.7.0",
]
